Revert "Feat. 프로토 타입 완성"

+++ src/main/java/com/tbp/honeyjar/admin/dto/notice/NoticeCorrectionRequestDto.java
@@ -7,7 +7,7 @@ import java.time.LocalDateTime;
 @Data
 public class NoticeCorrectionRequestDto {
 
-    private Long noticeId;
+    private int noticeId;
     private String title;
     private String post;
     private LocalDateTime updatedAt;
